meet/requirements.txt
Django==2.2.24
PyJWT==2.4.0
requests==2.25.1
simplejson==3.17.6
